# fix: improve getMetricName performance