Good day,

This PR improve the performance of `getMetricNameForBroker` and `getMetricNameForTopic` by removing the usage of `fmt`.

The [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) result are as follows. The benchmarks when run using `go test -bench '^\QBenchmark_getMetricNameFor' -run=^$ -count=10`.
```
goos: linux
goarch: amd64
pkg: github.com/IBM/sarama
cpu: AMD Ryzen 7 Pro 7735U with Radeon Graphics     
                           │ faec4c823ff893c7fc7b10021b66542fa77f7cb8.txt │ 07db3539fb81f1bf77d8a8a910b2b1e72b4c52c3.txt │
                           │                    sec/op                    │        sec/op         vs base                │
_getMetricNameForTopic-16                                    157.55n ± 0%            66.85n ± 1%  -57.57% (p=0.000 n=10)
_getMetricNameForBroker-16                                   108.50n ± 1%            43.12n ± 1%  -60.25% (p=0.000 n=10)
geomean                                                       130.7n                 53.69n       -58.93%

                           │ faec4c823ff893c7fc7b10021b66542fa77f7cb8.txt │ 07db3539fb81f1bf77d8a8a910b2b1e72b4c52c3.txt │
                           │                     B/op                     │         B/op          vs base                │
_getMetricNameForTopic-16                                      64.00 ± 0%             16.00 ± 0%  -75.00% (p=0.000 n=10)
_getMetricNameForBroker-16                                    28.000 ± 0%             4.000 ± 0%  -85.71% (p=0.000 n=10)
geomean                                                        42.33                  8.000       -81.10%

                           │ faec4c823ff893c7fc7b10021b66542fa77f7cb8.txt │ 07db3539fb81f1bf77d8a8a910b2b1e72b4c52c3.txt │
                           │                  allocs/op                   │      allocs/op        vs base                │
_getMetricNameForTopic-16                                      3.000 ± 0%             1.000 ± 0%  -66.67% (p=0.000 n=10)
_getMetricNameForBroker-16                                     2.000 ± 0%             1.000 ± 0%  -50.00% (p=0.000 n=10)
geomean                                                        2.449                  1.000       -59.18%
```

Thank you for reviewing this PR and let me know if anything needs to change.
Have a great day!