@dnwe 2025-05-13
Thanks @boekkooi-impossiblecloud – the change and benchmarked improvements look good to me, please can you amend your commits to add a DCO signoff to get that check to pass? See [here](https://github.com/IBM/sarama/blob/main/CONTRIBUTING.md#contributing-a-change)